Change how messages are stored in the mailbox and the information stored in the history
Mailbox now stores messages in maps of maps where only the order of messages between pairs of processes is kept. This is closer to what happens in a real erlang environment.
History now stores the possition of a received message in the queue of messages and the sender of that message.